Removed additional unused components.
